Rename worldTimeline to timelines.js
